requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
python-docx
openpyxl
PyMuPDF
pandas
Pillow
python-multipart
